+++ lib/cocoapods/validator.rb
@@ -345,8 +345,10 @@ module Pod
               next
             end
 
-            if message.include?('error: ')
+            if message =~ /\S+:\d+:\d+: error:/
               error('xcodebuild', message)
+            elsif message =~ /\S+:\d+:\d+: warning:/
+              warning('xcodebuild', message)
             else
               note('xcodebuild', message)
             end